fix clang-tidy readability-redundant-casting warnings
13 casts across 5 files that convert a value to the type it already has.
Most are (float) on RectF members, which are already float. The rest:
FileInfo::data is already char*, Size::dy is already int, and MobiDoc's
imageFirstRec is narrowed to int at the point it's assigned (line 550),
so casting it again does nothing.

Left the adjacent casts that are real conversions alone -
(float)(maxTileSize.dx + 1), (float)size.dx, (float)kThumbnailDx,
(int)d.UInt32BE() and (int)si.nPage all widen or narrow.

Verified by re-running clang-tidy over the five touched files: no
readability-redundant-casting left. Build is clean.

+++ src/MobiDoc.cpp
@@ -661,7 +661,7 @@ bool MobiDoc::DecodeExthHeader(const u8* data, int dataLen) {
             case 201:
                 if (length == 12 && imageFirstRec) {
                     d.Unskip(4);
-                    coverImageRec = (int)imageFirstRec + (int)d.UInt32BE();
+                    coverImageRec = imageFirstRec + (int)d.UInt32BE();
                 }
                 continue;
             case 503:
